update list homeStay by user

+++ src/components/Login/MyProperty.jsx
@@ -9,6 +9,55 @@ function MyProperty(props) {
     const [totalPages, setTotalPages] = useState(0);
     const [check, setCheck] = useState(false);
 
+    const visiblePages = totalPages;
+
+    const handlePageChange = (pageNumber) => {
+        setCurrentPage(pageNumber);
+        // Thực hiện các hành động khác khi chuyển trang
+        // ...
+    };
+
+    const renderPagination = () => {
+        const pageNumbers = [];
+        const halfVisiblePages = Math.floor(visiblePages / 2);
+        let startPage = currentPage - halfVisiblePages;
+        if (startPage < 0) startPage = 0;
+        let endPage = startPage + visiblePages - 1;
+        if (endPage > totalPages) {
+            endPage = totalPages;
+            startPage = endPage - visiblePages + 1;
+            if (startPage < 0) startPage = 0;
+        }
+
+        for (let i = startPage; i <= endPage; i++) {
+            const pageItemStyle = {
+                marginRight: '5px', // Khoảng cách giữa các số trang
+                display: 'inline-block', // Hiển thị trên cùng một dòng
+                cursor: 'pointer', // Con trỏ chuột thành dạng tay
+                fontWeight: currentPage === i ? 'bold' : 'normal', // Trang hiện tại được đậm
+            };
+            const pageLinkStyle = {
+                cursor: "pointer",
+                padding: '5px 10px', // Kích thước nút số trang
+                backgroundColor: currentPage === i ? '#ccc' : 'transparent', // Màu nền của trang hiện tại
+            };
+            pageNumbers.push(
+                <li key={i} style={pageItemStyle}>
+                    <button
+                        className="page-link"
+                        style={pageLinkStyle}
+                        onClick={() => handlePageChange(i)}
+                    >
+                        {i +1}
+                    </button>
+                </li>
+            );
+        }
+
+        return pageNumbers;
+    };
+
+
     useEffect(() => {
         const fetchData = async () => {
             try {
@@ -34,15 +83,18 @@ function MyProperty(props) {
     };
 
     const deleteHome = async (homeId) => {
-        try {
-            await axios.delete(`http://localhost:8080/homes/${homeId}`);
-            const updatedHomes = homes.filter((home) => home.id !== homeId);
-            setHomes(updatedHomes);
-        } catch (error) {
-            console.log(error);
+        const confirmed = window.confirm('Bạn chắc chắn muốn xóa?');
+        if (confirmed) {
+            try {
+                await axios.delete(`http://localhost:8080/homes/${homeId}`);
+                const updatedHomes = homes.filter((home) => home.id !== homeId);
+                setHomes(updatedHomes);
+            } catch (error) {
+                console.log(error);
+            }
         }
-
     };
+
 
     return (
         homes.length > 0 ? (
@@ -56,7 +108,7 @@ function MyProperty(props) {
                                     <tr>
                                         <td className="listing-photoo">
                                             <img alt="my-properties" src={home.image[0]}
-                                                 className="img-fluid"/>
+                                                 height={100}/>
                                         </td>
                                         <td className="title-container">
                                             <h5><a href="#">{home.name}</a></h5>
@@ -90,19 +142,20 @@ function MyProperty(props) {
                         </div>
                     ))}
                 </div>
-                <div className="pagination-container">
-                    <button
-                        onClick={goToPreviousPage}
-                        disabled={currentPage === 0}
+                <div className="pagination-container" style={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
+                    <button style={{border:"none", cursor:"pointer"}}
+                            onClick={goToPreviousPage}
+                            disabled={currentPage === 0}
                     >
-                        Previous
+                        <i style={{fontSize:25}} className="fa fa-angle-left"></i>
                     </button>
-                    <span>{currentPage + 1}</span> / <span>{totalPages}</span>
-                    <button
-                        onClick={goToNextPage}
-                        disabled={currentPage === totalPages - 1}
+                    {/*<span>{currentPage + 1}</span> / <span>{totalPages}</span>*/}
+                    {renderPagination()}
+                    <button style={{border:"none", cursor:"pointer"}}
+                            onClick={goToNextPage}
+                            disabled={currentPage === totalPages - 1}
                     >
-                        Next
+                        <i style={{fontSize:25}} className="fa fa-angle-right"></i>
                     </button>
                 </div>
             </div>
